+++ quick_blue/lib/src/messages.g.dart
@@ -926,6 +926,27 @@ class QuickBlueApi {
     );
   }
 
+  Future<List<String>> connectedDeviceIds(List<String> serviceUuids) async {
+    final pigeonVar_channelName =
+        'dev.flutter.pigeon.quick_blue.QuickBlueApi.connectedDeviceIds$pigeonVar_messageChannelSuffix';
+    final pigeonVar_channel = BasicMessageChannel<Object?>(
+      pigeonVar_channelName,
+      pigeonChannelCodec,
+      binaryMessenger: pigeonVar_binaryMessenger,
+    );
+    final Future<Object?> pigeonVar_sendFuture = pigeonVar_channel.send(
+      <Object?>[serviceUuids],
+    );
+    final pigeonVar_replyList = await pigeonVar_sendFuture as List<Object?>?;
+
+    final Object? pigeonVar_replyValue = _extractReplyValueOrThrow(
+      pigeonVar_replyList,
+      pigeonVar_channelName,
+      isNullValid: false,
+    );
+    return (pigeonVar_replyValue! as List<Object?>).cast<String>();
+  }
+
   Future<void> connect(String deviceId) async {
     final pigeonVar_channelName =
         'dev.flutter.pigeon.quick_blue.QuickBlueApi.connect$pigeonVar_messageChannelSuffix';