+++ quick_blue/lib/src/quick_blue_android.dart
@@ -73,6 +73,16 @@ class QuickBlueAndroid extends QuickBluePlatform {
   }
 
   @override
+  Future<List<BluetoothDevice>> connectedDevices({
+    List<String> serviceUuids = const <String>[],
+  }) async {
+    _ensureInitialized();
+
+    final deviceIds = await _api.connectedDeviceIds(serviceUuids);
+    return deviceIds.map(device).toList(growable: false);
+  }
+
+  @override
   Future<void> connect(String deviceId) {
     _ensureInitialized();