+++ quick_blue_windows/pigeons/messages.dart
@@ -108,6 +108,7 @@ abstract class QuickBlueApi {
     Map<int, Uint8List>? manufacturerData,
   });
   void stopScan();
+  List<String> connectedDeviceIds(List<String> serviceUuids);
   void connect(String deviceId);
   void disconnect(String deviceId);